Updated dependencies and added client validation.
Added references to new packages such as `FluentValidation`, `Google.Protobuf` and `Grpc.Tools` in the `Connectiq.Contracts.csproj` file, along with a Protobuf file.

In `Program.cs`, the `MassTransit` configuration was removed and services for validators and infrastructure were added.

A new validator was created in `CustomerValidator.cs` to validate properties of the `Customer` object, ensuring that the required fields are not empty and comply with the specified formats.

In addition, methods were implemented in `ServiceCollectionExtensions.cs` to register validators and configure `MassTransit` with RabbitMQ.

Translated with DeepL.com (free version)